Changed name of Header Identity filter to Header User
