+++ repose-aggregator/artifacts/filter-bundle/build.gradle
@@ -10,9 +10,9 @@ dependencies {
     earlib project(":repose-aggregator/components/filters/flush-output-filter")
     earlib project(":repose-aggregator/components/filters/forwarded-proto-filter")
     earlib project(":repose-aggregator/components/filters/header-id-mapping-filter")
-    earlib project(":repose-aggregator/components/filters/header-identity")
     earlib project(":repose-aggregator/components/filters/header-normalization-filter")
     earlib project(":repose-aggregator/components/filters/header-translation-filter")
+    earlib project(":repose-aggregator/components/filters/header-user-filter")
     earlib project(":repose-aggregator/components/filters/herp-filter")
     earlib project(":repose-aggregator/components/filters/ip-user-filter")
     earlib project(":repose-aggregator/components/filters/iri-validator-filter")
