--- repose-aggregator/components/filters/header-identity/pom.xml
@@ -1,114 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>org.openrepose</groupId>
-        <artifactId>filters-support</artifactId>
-        <version>8.0.0.0-SNAPSHOT</version>
-    </parent>
-
-    <name>Repose Components - Client Identity By Header</name>
-    <artifactId>header-identity</artifactId>
-    <packaging>jar</packaging>
-
-    <description>
-        TODO
-    </description>
-
-    <properties>
-        <sonar.jacoco.itReportPath>${project.basedir}/../../../target/jacoco-it.exec</sonar.jacoco.itReportPath>
-    </properties>
-
-    <dependencies>
-        <!-- servlet filter stuff -->
-        <dependency>
-            <groupId>org.openrepose</groupId>
-            <artifactId>core-service-api</artifactId>
-        </dependency>
-
-        <!-- JAXB toString -->
-        <dependency>
-            <groupId>org.jvnet.jaxb2_commons</groupId>
-            <artifactId>jaxb2-basics-runtime</artifactId>
-        </dependency>
-
-        <!-- HTTP request/response wrappers -->
-        <dependency>
-            <groupId>org.openrepose</groupId>
-            <artifactId>commons-utilities</artifactId>
-        </dependency>
-
-        <!-- Let's use Scala -->
-        <dependency>
-            <groupId>org.scala-lang</groupId>
-            <artifactId>scala-library</artifactId>
-        </dependency>
-
-        <!-- Scalatest for testing scala -->
-        <dependency>
-            <groupId>org.scalatest</groupId>
-            <artifactId>scalatest_2.10</artifactId>
-        </dependency>
-
-        <!-- Lazy logging -->
-        <dependency>
-            <groupId>com.typesafe.scala-logging</groupId>
-            <artifactId>scala-logging-slf4j_2.10</artifactId>
-            <version>2.1.2</version>
-        </dependency>
-
-        <!-- MockHttpServletRequest/Response -->
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-test</artifactId>
-            <scope>test</scope>
-        </dependency>
-
-        <!-- LogFactory, needed by MockHttpServletRequest/Response -->
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>jcl-over-slf4j</artifactId>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.codehaus.mojo</groupId>
-                <artifactId>xml-maven-plugin</artifactId>
-
-                <configuration>
-                    <transformationSets>
-                        <transformationSet>
-                            <dir>src/main/resources/META-INF/schema/config</dir>
-                            <stylesheet>../xsl/remove-1.1-elements.xsl</stylesheet>
-                            <outputDir>${project.build.directory}/generated-resources/xml/xslt/config</outputDir>
-                        </transformationSet>
-                    </transformationSets>
-                </configuration>
-            </plugin>
-
-            <plugin>
-                <groupId>org.jvnet.jaxb2.maven2</groupId>
-                <artifactId>maven-jaxb2-plugin</artifactId>
-
-                <configuration>
-
-                    <specVersion>2.2</specVersion>
-                    <schemaDirectory>${project.build.directory}/generated-resources/xml/xslt</schemaDirectory>
-
-                    <schemaIncludes>
-                        <include>**/*.xsd</include>
-                    </schemaIncludes>
-                    <bindingIncludes>
-                        <include>**/*.xjb</include>
-                    </bindingIncludes>
-                    <strict>true</strict>
-                    <verbose>false</verbose>
-
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-</project>
