--- repose-aggregator/components/filters/header-identity/src/main/resources/META-INF/schema/examples/header-identity.cfg.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<header-identity xmlns='http://docs.openrepose.org/repose/header-identity/v1.0'>
-
-    <source-headers>
-        <header id="X-Auth-Token"/>
-        <header id="X-Forwarded-For" quality="0.2"/>
-    </source-headers>
-
-</header-identity>
