+++ repose-aggregator/components/filters/pom.xml
@@ -17,7 +17,7 @@
 
     <modules>
         <module>slf4j-http-logging-filter</module>
-        <module>header-identity</module>
+        <module>header-user-filter</module>
         <module>header-id-mapping-filter</module>
         <module>uri-user-filter</module>
         <module>rackspace-auth-user-filter</module>
