--- repose-aggregator/functional-tests/spock-functional-test/src/test/configs/features/core/configloadingandreloading/header-identity-bad/header-identity.cfg.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<header-identity xmlns:xsi='http://www.w3.org/2001/XMLSchema-instance'
-                 xmlns='http://docs.openrepose.org/repose/header-identity/v1.0'
-                 xsi:schemaLocation='http://docs.openrepose.org/repose/header-identity/v1.0 ../config/header-identity-configuration.xsd'>
-
-    <source-headers>
-        <header id="X-Forwarded-For" quality="0.2"/>
-    </source-headers>
-
