--- repose-aggregator/functional-tests/spock-functional-test/src/test/configs/features/core/configloadingandreloading/header-identity-common/system-model.cfg.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<system-model xmlns="http://docs.openrepose.org/repose/system-model/v2.0">
-    <repose-cluster id="repose">
-
-        <nodes>
-            <node id="config-test" hostname="localhost" http-port="${reposePort}"/>
-        </nodes>
-
-        <filters>
-            <filter name="header-identity"/>
-        </filters>
-
-        <destinations>
-            <endpoint id="mock-service" protocol="http" hostname="${targetHostname}" port="${targetPort}" root-path="/"
-                      default="true"/>
-        </destinations>
-
-    </repose-cluster>
-</system-model>
