+++ repose-aggregator/functional-tests/spock-functional-test/src/test/configs/features/core/configloadingandreloading/header-user-good/header-user.cfg.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<header-user xmlns:xsi='http://www.w3.org/2001/XMLSchema-instance'
+                 xmlns='http://docs.openrepose.org/repose/header-user/v1.0'
+                 xsi:schemaLocation='http://docs.openrepose.org/repose/header-user/v1.0 ../config/header-user-configuration.xsd'>
+
+    <source-headers>
+        <header id="X-Forwarded-For" quality="0.2"/>
+    </source-headers>
+
+</header-user>
