--- repose-aggregator/functional-tests/spock-functional-test/src/test/configs/features/core/intrafilterlogging/header-identity.cfg.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<header-identity xmlns='http://docs.openrepose.org/repose/header-identity/v1.0'>
-
-    <source-headers>
-        <header id="X-Header-User" quality="0.2"/>
-    </source-headers>
-</header-identity>
