+++ repose-aggregator/functional-tests/spock-functional-test/src/test/configs/features/core/intrafilterlogging/header-user.cfg.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<header-user xmlns='http://docs.openrepose.org/repose/header-user/v1.0'>
+
+    <source-headers>
+        <header id="X-Header-User" quality="0.2"/>
+    </source-headers>
+</header-user>
