+++ repose-aggregator/functional-tests/spock-functional-test/src/test/configs/features/core/intrafilterlogging/system-model.cfg.xml
@@ -14,7 +14,7 @@
             <filter name="openstack-identity-v3" />
             <filter name="ip-user" />
             <filter name="content-type-stripper"/>
-            <filter name="header-identity"/>
+            <filter name="header-user"/>
             <filter name="header-normalization"/>
             <filter name="header-translation"/>
             <filter name="header-id-mapping" />
