+++ repose-aggregator/functional-tests/spock-functional-test/src/test/configs/features/core/startup/system-model.cfg.xml
@@ -11,7 +11,7 @@
             <filter name="ip-user"/>
             <filter name="destination-router"/>
             <filter name="header-id-mapping"/>
-            <filter name="header-identity"/>
+            <filter name="header-user"/>
             <filter name="header-normalization"/>
             <filter name="header-translation"/>
             <filter name="rate-limiting"/>
