--- repose-aggregator/functional-tests/spock-functional-test/src/test/configs/features/filters/headeridentity/system-model.cfg.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<system-model xmlns="http://docs.openrepose.org/repose/system-model/v2.0">
-    <repose-cluster id="repose">
-
-        <nodes>
-            <node id="node1" hostname="localhost" http-port="${reposePort}"/>
-        </nodes>
-
-        <filters>
-            <filter name="header-identity"/>
-        </filters>
-
-        <destinations>
-            <endpoint id="openrepose" protocol="http" hostname="localhost" root-path="/" port="${targetPort}"
-                      default="true"/>
-        </destinations>
-
-    </repose-cluster>
-
-</system-model>
-
