+++ repose-aggregator/functional-tests/spock-functional-test/src/test/groovy/features/core/configloadingandreloading/StartWithBadConfigsTest.groovy
@@ -96,7 +96,7 @@ class StartWithBadConfigsTest extends Specification {
         "keystone-v2"             | _
         "dist-datastore"          | _
         "uri-user"                | _
-        "header-identity"         | _
+        "header-user"             | _
         "header-id-mapping"       | _
         "ip-user"                 | _
         "validator"               | _
