+++ repose-aggregator/functional-tests/spock-functional-test/src/test/groovy/features/core/configloadingandreloading/TransitionBadToGoodConfigsTest.groovy
@@ -118,7 +118,7 @@ class TransitionBadToGoodConfigsTest extends Specification {
         "keystone-v2"             | 200
         "dist-datastore"          | 200
         "uri-user"                | 200
-        "header-identity"         | 200
+        "header-user"             | 200
         "header-id-mapping"       | 200
         "ip-user"                 | 200
         "validator"               | 200
