+++ repose-aggregator/functional-tests/spock-functional-test/src/test/groovy/features/core/intrafilterlogging/IntraFilterLoggingTest.groovy
@@ -79,7 +79,7 @@ class IntraFilterLoggingTest extends ReposeValveTest {
         reposeLogSearch.searchByString("\"currentFilter\":\"add-header\"").size() > 0
         reposeLogSearch.searchByString("\"currentFilter\":\"keystone-v2\"").size() > 0
         reposeLogSearch.searchByString("\"currentFilter\":\"openstack-identity-v3\"").size() > 0
-        reposeLogSearch.searchByString("\"currentFilter\":\"header-identity\"").size() > 0
+        reposeLogSearch.searchByString("\"currentFilter\":\"header-user\"").size() > 0
         reposeLogSearch.searchByString("\"currentFilter\":\"ip-user\"").size() > 0
         reposeLogSearch.searchByString("\"currentFilter\":\"header-normalization\"").size() > 0
         reposeLogSearch.searchByString("\"currentFilter\":\"header-translation\"").size() > 0
@@ -101,7 +101,7 @@ class IntraFilterLoggingTest extends ReposeValveTest {
         reposeLogSearch.searchByString("null-add-header").size() == 0
         reposeLogSearch.searchByString("null-keystone-v2").size() == 0
         reposeLogSearch.searchByString("null-openstack-identity-v3").size() == 0
-        reposeLogSearch.searchByString("null-header-identity").size() == 0
+        reposeLogSearch.searchByString("null-header-user").size() == 0
         reposeLogSearch.searchByString("null-ip-user").size() == 0
         reposeLogSearch.searchByString("null-header-normalization").size() == 0
         reposeLogSearch.searchByString("null-header-translation").size() == 0
