+++ repose-aggregator/installation/configs/filters/header-user.cfg.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!-- http://wiki.openrepose.org/display/REPOSE/Header+Identity -->
+<header-user xmlns:xsi='http://www.w3.org/2001/XMLSchema-instance'
+                 xmlns='http://docs.openrepose.org/repose/header-user/v1.0'
+                 xsi:schemaLocation='http://docs.openrepose.org/repose/header-user/v1.0 ../config/header-user-configuration.xsd'>
+
+    <source-headers>
+        <header id="X-Auth-Token" quality=".95"/>
+        <header id="X-Forwarded-For" quality="0.5"/>
+    </source-headers>
+
+</header-user>
