+++ repose-aggregator/installation/list-repose-files.sh
@@ -32,7 +32,7 @@ FILES=(\
 /etc/repose/destination-router.cfg.xml \
 #/etc/repose/dist-datastore.cfg.xml \
 /etc/repose/fail-404.wadl \
-/etc/repose/header-identity.cfg.xml \
+/etc/repose/header-user.cfg.xml \
 /etc/repose/header-id-mapping.cfg.xml \
 /etc/repose/header-normalization.cfg.xml \
 /etc/repose/header-translation.cfg.xml \
