+++ settings.gradle
@@ -20,9 +20,9 @@ include 'repose-aggregator',
         'repose-aggregator/components/filters/flush-output-filter',
         'repose-aggregator/components/filters/forwarded-proto-filter',
         'repose-aggregator/components/filters/header-id-mapping-filter',
-        'repose-aggregator/components/filters/header-identity',
         'repose-aggregator/components/filters/header-normalization-filter',
         'repose-aggregator/components/filters/header-translation-filter',
+        'repose-aggregator/components/filters/header-user-filter',
         'repose-aggregator/components/filters/herp-filter',
         'repose-aggregator/components/filters/ip-user-filter',
         'repose-aggregator/components/filters/iri-validator-filter',
